// src/components/ui/ranking/Filter.module.css.d.ts
declare const styles: Readonly<{ filter: string }> &
	Readonly<{ button: string }> &
	Readonly<{ button: string }> &
	Readonly<{ button: string }> &
	Readonly<{ label: string }> &
	Readonly<{ label: string }> &
	Readonly<{ icon: string }> &
	Readonly<{ open: string }> &
	Readonly<{ filter_transition: string }> &
	Readonly<{ filter_transition: string }> &
	Readonly<{ filter_transition: string }> &
	Readonly<{ filter_transition: string }> &
	Readonly<{ filter_transition: string }> &
	Readonly<{ filter_transition: string }> &
	Readonly<{ filter_transition: string }> &
	Readonly<{ enter: string }> &
	Readonly<{ enter_from: string }> &
	Readonly<{ enter_to: string }> &
	Readonly<{ leave: string }> &
	Readonly<{ leave_from: string }> &
	Readonly<{ leave_to: string }> &
	Readonly<{ panel: string }> &
	Readonly<{ container: string }> &
	Readonly<{ container: string }>;
export default styles;
//# sourceMappingURL=./Filter.module.css.d.ts.map